break up alternative solutions into their own views
